Primera corrección de GPU integrate
