+++ route/a_homePage.py
@@ -1,0 +1,7 @@
+from django.shortcuts import render
+
+#home page
+def home(request):
+	#Return page	
+	return render(request, 'home.html')
+
