+++ route/templates/home.html
@@ -54,7 +54,7 @@
                 </div>
                 <div class="cardParameters">
                     <div class="firstElementParameters">
-                       <p class="firstElementText">StreetSelection</p>      
+                       <p class="firstElementText">Street selection</p>      
                         {% load static %}
                         <img src="{% static "route/images/cruceroOpenStreetMap.png" %}" alt ="Crucero" class="imageCrucero"/>
                     </div>
