+++ simulatorver03/urls.py
@@ -16,41 +16,58 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path
 
+#Home view
+from route import a_homePage as home
 ##Route views
-from route import generalMap 
-from route import viewsPythonArray as pythonArray_views
-from route import viewsGPU as GPU_views
-from route import testIntegration as testIntegration_views
-from route import viewsGPUIntegrateGPU as integrateGPU
-from route import homePage as home
+from route import a_generalMap 
+##CPU classes
+from route import CPUClasses
+##CPU arreglos
+from route import CPUArrays
 
+#GPU simulator, integration CPU
+from route import GPU_NI
+from route import GPU_testGama
+from route import GPU_integrate
+
+#General
 urlpatterns = [
+    #GENERAL
+    #HOME
+    path('', home.home, name ='home'),
+    
+    #Parse info
+    path('parse/', a_generalMap.parse), # Only to get info from nodes and ways in a vehicular network
 
     #General maps
-    path('',generalMap.showmap,name='showmap'), #Muestra un mapa random
+    path('showmapRandom',a_generalMap.showmapRandom,name='showmapRandom'), #Muestra un mapa random
     path('<str:lat1>,<str:long1>,<str:lat2>,<str:long2>',
-        generalMap.showroute,name='showroute'), #Muestra una ruta del punto A al B
-    path('showmapSimulator/',
-     generalMap.showmapSimulator), # Pinta la ruta de un punto A a un punto B del Centro histórico de la Ciudad de
-    path('showmapInitial/<int:selector>', generalMap.showmapInitial),
-    
-    path('testGPU1/', generalMap.testGPU1),
-    path('testGPU2/', generalMap.testGPU2),
-    path('CPUClasses/', generalMap.CPUClasses),
-    path('CPUArray/', pythonArray_views.pythonArray),
-    path('GPU/', GPU_views.pythonGPU),
-    #path('IntegrateGPU/', GPU_views.pythonGPUIntegrate),
-    path('TestIntegrationGPU/', testIntegration_views.testIntegrationGPU, name="gpuGenericIntegration"), #Generic GPU integration code 
-    path('IntegrateGPU/', integrateGPU.pythonGPUIntegrate), #GPU Integration to the model
+        a_generalMap.showRoute,name='showroute'), #Muestra una ruta del mapa
+    path('showmapMexicoCityPoints/',
+     a_generalMap.showmapMexicoCityPoints), # Pinta la ruta de un punto A a un punto B del Centro histórico de la Ciudad de
+    path('showmapSelector/<int:selector>', a_generalMap.showmapSelector),
 
-    ##Admin
+    #ADMIN
     path('admin/', admin.site.urls), #admin to check models 
-    path('simulator03/', generalMap.parse), # Only to get info from nodes and ways in a vehicular network
-
+    #END OF GENERAL
 ]
-
-#To add html views
-
+#CPU
 urlpatterns += [
-    path('home/', home.home, name ='home')
+    #CPU
+    #Classes
+    path('CPUClasses/', CPUClasses.CPUClasses),
+    #Arrays
+    path('CPUArrays/', CPUArrays.CPUArrays),
+    ]
+#GPU
+urlpatterns += [
+   #GPU
+    #without integration method
+    path('GPU_NI/', GPU_NI.GPU_NI),
+    #test of generic 
+    path('GPU_testGeneric/', GPU_testGama.GPU_testGeneric, name="GPU_testGeneric"), #some generic tests
+    #test of integration generic Dr. Gamaliel working on GPU
+    path('GPU_testGama/', GPU_testGama.GPU_testGama, name="GPU_testGama"), #Generic GPU integration code Dr. Gamaliel
+    #with integration method
+    path('GPU_integrate/', GPU_integrate.GPU_integrate), #GPU Integration to the model
 ]
